Housekeeping PR that does the following: 1. removes component level READMEs. 2. ensures all ldpath props are prefaced by "ldpath." 3. renames fcrepo-indexing-http  to fcrepo-http-forwarder in order to underscore that it is not necessarily just for indexing. 4. brings other properties into alignment with the convention of ensuring that all service props are prefixed by the service name.
