+++ docker-compose/camel-toolbox-config/configuration.properties
@@ -3,10 +3,10 @@ fcrepo.authHost=fcrepo
 
 jms.brokerUrl=tcp://fcrepo:61616
 
-solr.indexer.enabled=true
+solr.indexing.enabled=true
 solr.baseUrl=http://solr:8983/solr/fcrepo
 
-triplestore.indexer.enabled=true
+triplestore.indexing.enabled=true
 triplestore.baseUrl=http://fuseki:3030/fcrepo
 
 reindexing.rest.host=0.0.0.0
