+++ fcrepo-camel-toolbox-app/pom.xml
@@ -34,7 +34,7 @@
 
     <dependency>
       <groupId>${project.parent.groupId}</groupId>
-      <artifactId>fcrepo-indexing-http</artifactId>
+      <artifactId>fcrepo-http-forwarding</artifactId>
       <version>${project.parent.version}</version>
     </dependency>
