+++ fcrepo-http-forwarding/src/main/resources/org/fcrepo/camel/httpforwarding/httpMessage.mustache
