+++ fcrepo-ldpath/src/test/java/org/fcrepo/camel/ldpath/RouteTest.java
@@ -80,10 +80,10 @@ public class RouteTest {
         final var ldCacheDir = targetDir + File.separator + RouteTest.class.getCanonicalName() +
                 File.separator + "ldcache";
         new File(ldCacheDir).mkdirs();
-        System.setProperty("ldcache.directory", ldCacheDir);
+        System.setProperty("ldpath.ldcache.directory", ldCacheDir);
         final String restPort = System.getProperty("fcrepo.dynamic.ldpath.port", "9085");
-        System.setProperty("rest.port", restPort);
-        System.setProperty("rest.host", "0.0.0.0");
+        System.setProperty("ldpath.rest.port", restPort);
+        System.setProperty("ldpath.rest.host", "0.0.0.0");
     }
 
     @AfterClass
